updated ui/pattern-ls2request  to use redux toolkit
