+++ ui/pattern-ls2request/src/App/App.js
@@ -1,20 +1,16 @@
 import Button from '@enact/ui/Button';
 import Item from '@enact/ui/Item';
 import {useCallback, useEffect} from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-// import {connect} from 'react-redux';
+import {useSelector, useDispatch} from 'react-redux';
 
-import {getSystemSettings, setSystemSettings, setSystemSettingsSubscribed} from '../actions';
+import {getSystemSettings, setSystemSettings, setSystemSettingsSubscribed} from '../store';
 
 import css from './App.module.less';
 
 const App = () => {
-
-	// const {brightness, eyeComfortMode} = useSelector(state => {
-	// 	console.log(state)
-	// });
-	const brightness = '100', eyeComfortMode = '0';
+	const {brightness, eyeComfortMode} = useSelector(state => state);
 	const dispatch = useDispatch();
+
 	useEffect(() => {
 		if (typeof window !== 'undefined' && typeof window.PalmSystem !== 'undefined') {
 			dispatch(getSystemSettings({
@@ -30,6 +26,7 @@ const App = () => {
 			}));
 		}
 	}, []); // eslint-disable-line react-hooks/exhaustive-deps
+
 	const handleDecreaseBrightness = useCallback(() => {
 		let brightnessDec = Number(brightness);
 		brightnessDec = brightnessDec !== 0 ? brightnessDec - 10 : brightnessDec;
@@ -40,6 +37,7 @@ const App = () => {
 			}
 		}));
 	}, [brightness, dispatch]);
+
 	const handleIncreaseBrightness = useCallback(() => {
 		let brightnessInc = Number(brightness);
 		brightnessInc = brightnessInc !== 100 ? brightnessInc + 10 : brightnessInc;
@@ -50,6 +48,7 @@ const App = () => {
 			}
 		}));
 	}, [brightness, dispatch]);
+
 	// if subscribed, we don't need to invoke redux chain as subscribed instance will invoke the data flow
 	const onEyeComfortModeToggle = useCallback(() => setSystemSettingsSubscribed({
 		category: 'picture',
@@ -57,19 +56,20 @@ const App = () => {
 			'eyeComfortMode': eyeComfortMode === 'on' ? 'off' : 'on'
 		}
 	}), [eyeComfortMode]);
+
 	const checkSystem = () => {
 		if (typeof window === 'undefined' || typeof window.PalmSystem === 'undefined') {
 			return <div>This test will only function correctly on webOS systems!</div>;
 		}
 	};
 
-	// if (typeof window === 'undefined' || typeof window.PalmSystem === 'undefined') {
-	// 	return <div className={css.main}>This test will only function correctly on webOS systems!</div>;
-	// }
+	if (typeof window === 'undefined' || typeof window.PalmSystem === 'undefined') {
+		return <div className={css.main}>This test will only function correctly on webOS systems!</div>;
+	}
 
 	return (
 		<div className={css.main}>
-			{/* {checkSystem()} */}
+			{checkSystem()}
 			<Item>
 				Brightness : {brightness}
 			</Item>
@@ -83,19 +83,5 @@ const App = () => {
 	);
 
 };
-
-// App.propTypes = {
-// 	dispatch: PropTypes.func.isRequired,
-// 	brightness: PropTypes.string,
-// 	eyeComfortMode: PropTypes.string
-// };
-
-// const mapStateToProps = ({systemSettings}) => {
-// 	const {brightness, eyeComfortMode} = systemSettings;
-// 	return {
-// 		brightness,
-// 		eyeComfortMode
-// 	};
-// };
 
 export default (App);
