--- ui/pattern-ls2request/src/actions/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "actions.js"
-}
