--- ui/pattern-ls2request/src/reducers/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "reducers.js"
-}
